react-poc/leasingborsen-react-production/archive/projects/railway-pdfplumber-poc/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pdfplumber==0.10.3
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.4
aiohttp==3.9.1
pydantic==2.5.3
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.25.2